backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlmodel>=0.0.19
sqlalchemy>=2.0.32
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pymysql>=1.1.0
cryptography>=41.0.0
python-multipart>=0.0.6
bcrypt>=4.0.0